feat(User): add logOut function to handle user sign-out functionality using Firebase Authentication library
The logOut.jsx file contains the implementation of the logOut function. This function utilizes the Firebase Authentication library to perform a sign-out operation. It imports the getAuth and signOut functions from the firebase/auth module.

The logOut function initializes the auth variable using the getAuth function. It then calls the signOut function, passing in the auth variable, to initiate the sign-out process. If the sign-out is successful, the function does nothing. If an error occurs during the sign-out process, an error is thrown.

This logOut function can be used to handle user sign-out functionality in the application.