refactor(App.jsx): optimize imports and use lazy loading for components
The changes in this commit optimize the imports in the `App.jsx` file by using relative paths and the `@features` alias. This improves the readability and maintainability of the code.

Additionally, the code now uses the `lazy` function from React to dynamically import the components `Products`, `Product`, `Login`, and `Home`. This improves the performance of the application by only loading these components when they are actually needed.

The `Suspense` component is also added to render a loading spinner with the class name "lds-circle" until the lazy-loaded components are fully loaded and ready to be rendered. This provides a better user experience by indicating that the content is being loaded.